move project update to workspace
