+++ lib/api/v3/workspaces/nested_apis.rb
@@ -33,6 +33,7 @@ module API
     module Workspaces
       class NestedApis < ::API::OpenProjectAPI
         mount ::API::V3::Projects::UpdateFormAPI
+        patch &::API::V3::Utilities::Endpoints::Update.new(model: Project).mount
 
         mount API::V3::Workspaces::AvailableAssigneesAPI
         mount API::V3::Types::TypesByWorkspaceAPI
