+++ spec/lib/api/v3/projects/project_representer_rendering_spec.rb
@@ -630,7 +630,7 @@ RSpec.describe API::V3::Projects::ProjectRepresenter, "rendering" do
 
         it_behaves_like "has an untitled link" do
           let(:link) { "updateImmediately" }
-          let(:href) { api_v3_paths.project project.id }
+          let(:href) { api_v3_paths.workspace project.id }
         end
       end
